refactor: change packaging
