+++ src/test/kotlin/racingcar/RoundTest.kt
@@ -5,6 +5,7 @@ import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.ShouldSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
+import racingcar.game.Round
 
 class RoundTest : ShouldSpec({
     context("Create") {
